front: rollingstockeditor: isolate search filters from rollingstockmodal to use it in rollingstockeditor
- refacto: add missing length, mass and max_speed

- fix comments
