+++ front/public/locales/en/rollingStockEditor.json
@@ -1,24 +1,51 @@
 {
   "errors": {
-    "missingInformation": "Missing information",
-    "must be number": "Number is required",
-    "must be equal to one of the allowed values": "Authorised value required"
+    "missingInformation": "missing informations",
+    "must be number": "must be a number",
+    "must be equal to one of the allowed values": "must be equal to one of the allowed values"
   },
-  "startup_time": "Releasing time (s)",
-  "startup_acceleration": "Starting acceleration (m/s²)",
-  "comfort_acceleration": "Acceleration for comfort (m/s²)",
-  "inertia_coefficient": "Factor of inertia",
-  "gamma_value": "Timetable deceleration (m/s²)",
-  "loading_gauge": "Gauge",
-  "rollingResistance": "Resistance to movement",
-  "rollingResistanceFormula": "(according to Davis's formule, R(v) = a + bv + cv²)",
-  "rolling_resistance_A": "Solid friction (N)",
-  "rolling_resistance_B": "Viscous friction (N.s/m)",
-  "rolling_resistance_C": "Aerodynamic friction (N.s²/m²)",
-  "selectRollingStock": "Select a rolling stock to display informations",
-  "confirmAction": "Would you like to confirm the change?",
-  "cancelAction": "Would you like to cancel the modification?",
+  "messages": {
+    "success": "success",
+    "failure": "failure",
+    "rollingStockAdded": "the rolling stock has been added.",
+    "rollingStockUpdated": "the rolling stock has been updated.",
+    "rollingStockDeleted": "the rolling stock has been deleted.",
+    "rollingStockNotAdded": "the rolling stock has not been added.",
+    "rollingStockNotUpdated": "the rolling stock has not been updated.",
+    "rollingStockNotDeleted": "the rolling stock has not been deleted."
+  },
+  "name": "Name",
+  "version": "Version",
+  "detail": "Detail",
+  "family": "Family",
+  "grouping": "Grouping",
+  "number": "Number",
+  "reference": "Reference",
+  "series": "Series",
+  "subseries": "Subseries",
+  "type": "Type",
+  "unit": "Unit",
+  "length": "Length",
+  "mass": "Mass",
+  "max_speed": "Max speed",
+  "startup_time": "Startup time",
+  "startup_acceleration": "Acceleration « startup »",
+  "comfort_acceleration": "Acceleration « comfort »",
+  "inertia_coefficient": "Inertia coefficient",
+  "gamma_value": "Gamma value",
+  "loading_gauge": "Loading gauge",
+  "rollingResistance": "Rolling resistance",
+  "rollingResistanceFormula": "(in accordance with the Davis formula, R(v) = a + bv + cv²)",
+  "rolling_resistance_A": "Frottements solides",
+  "rolling_resistance_B": "viscous friction",
+  "rolling_resistance_C": "aerodynamic resistance",
+  "selectRollingStock": "Select a rolling stock to display its information.",
+  "addNewRollingStock": "New rolling stock",
+  "confirmAction": "Do you want to confirm this action?",
+  "cancelAction": "Do you want to cancel this action?",
   "back": "Back",
   "cancel": "Cancel",
-  "confirm": "Confirm"
+  "confirm": "Confirm",
+  "waitingLoader": "Please wait, loading rolling stock database...",
+  "noResultFound": "No result"
 }
