+++ front/public/locales/fr/rollingStockEditor.json
@@ -45,5 +45,7 @@
   "cancelAction": "Souhaitez-vous annuler cette action ?",
   "back": "Retour",
   "cancel": "Annuler",
-  "confirm": "Confirmer"
+  "confirm": "Confirmer",
+  "waitingLoader": "Veuillez patienter, chargement de la base de données matériel…",
+  "noResultFound": "Aucun résultat"
 }
