+++ front/src/applications/rollingStockEditor/json/RollingStockBaseOpenapiSchema.json
@@ -114,6 +114,9 @@
           "subseries",
           "type",
           "unit",
+          "length",
+          "mass",
+          "max_speed",
           "startup_time",
           "startup_acceleration",
           "comfort_acceleration",
