+++ front/src/applications/rollingStockEditor/views/RollingStockEditor.tsx
@@ -6,6 +6,8 @@ import { getRollingStockID } from 'reducers/osrdconf/selectors';
 import Loader from 'common/Loader';
 import { useTranslation } from 'react-i18next';
 import { LightRollingStock } from 'common/api/osrdEditoastApi';
+import SearchRollingStock from 'common/RollingStockSelector/SearchRollingStock';
+import { isEmpty } from 'lodash';
 import RollingStockEditorButtons from '../components/RollingStockEditorButtons';
 import RollingStockEditorCard from '../components/RollingStockEditorCard';
 import RollingStockEditorForm from '../components/RollingStockEditorForm';
@@ -18,6 +20,10 @@ export default function RollingStockEditor({ rollingStocks }: RollingStockEditor
   const { t } = useTranslation('rollingStockEditor');
   const ref2scroll: React.RefObject<HTMLInputElement> = useRef<HTMLInputElement>(null);
   const rollingStockID = useSelector(getRollingStockID);
+  const [filteredRollingStockList, setFilteredRollingStockList] =
+    useState<LightRollingStock[]>(rollingStocks);
+  const [isLoading, setIsLoading] = useState(false);
+
   const [isEditing, setIsEditing] = useState(false);
   const [isAdding, setIsAdding] = useState(false);
 
@@ -25,8 +31,77 @@ export default function RollingStockEditor({ rollingStocks }: RollingStockEditor
 
   const selectedRollingStock = rollingStocks.find((rs) => rs.id === openedRollingStockCardId);
 
+  const listOfRollingStocks = (
+    <div className="rollingstock-editor-list pr-1">
+      {filteredRollingStockList.length > 0 &&
+        filteredRollingStockList.map((data) => (
+          <div key={data.id}>
+            <div className="d-flex">
+              <div
+                role="button"
+                tabIndex={-1}
+                className="d-flex align-self-start rollingstock-elements w-100"
+                onClick={() => {
+                  setIsEditing(false);
+                  setIsAdding(false);
+                }}
+              >
+                <RollingStockCard
+                  isOnEditMode
+                  rollingStock={data}
+                  noCardSelected={openedRollingStockCardId === undefined}
+                  isOpen={data.id === openedRollingStockCardId}
+                  setOpenedRollingStockCardId={setOpenedRollingStockCardId}
+                  ref2scroll={rollingStockID === data.id ? ref2scroll : undefined}
+                />
+              </div>
+              {data.id === openedRollingStockCardId && (
+                <div className="align-self-start">
+                  <RollingStockEditorButtons
+                    isCondensed
+                    rollingStock={data}
+                    setIsEditing={setIsEditing}
+                    isRollingStockLocked={data.locked as boolean}
+                  />
+                </div>
+              )}
+            </div>
+            {openedRollingStockCardId === data.id && (
+              <div className="d-flex flex-column pl-0 rollingstock-editor-form-container mb-3">
+                {(selectedRollingStock || isEditing) &&
+                  ((selectedRollingStock && !isEditing && (
+                    <RollingStockEditorCard
+                      isEditing={isEditing}
+                      data={selectedRollingStock}
+                      setIsEditing={setIsEditing}
+                    />
+                  )) ||
+                    (isEditing && (
+                      <RollingStockEditorForm
+                        rollingStockData={selectedRollingStock}
+                        setIsEditing={setIsEditing}
+                        setIsAdding={setIsAdding}
+                      />
+                    )))}
+              </div>
+            )}
+          </div>
+        ))}
+    </div>
+  );
+
+  function displayList() {
+    if (isEmpty(filteredRollingStockList)) {
+      if (isLoading) {
+        return <Loader msg={t('waitingLoader')} />;
+      }
+      return <div className="rollingstock-empty mx-auto">{t('noResultFound')}</div>;
+    }
+    return listOfRollingStocks;
+  }
+
   return (
-    <div className="d-flex pt-5 mt-5">
+    <div className="d-flex rollingstock-editor">
       <div className="d-flex ml-4 flex-column rollingstock-editor-left-container">
         <div className="d-flex justify-content-center w-100">
           <button
@@ -41,6 +116,15 @@ export default function RollingStockEditor({ rollingStocks }: RollingStockEditor
             {t('addNewRollingStock')}
           </button>
         </div>
+        <SearchRollingStock
+          rollingStocks={rollingStocks}
+          rollingStockID={openedRollingStockCardId}
+          setOpenedRollingStockCardId={setOpenedRollingStockCardId}
+          setFilteredRollingStockList={setFilteredRollingStockList}
+          filteredRollingStockList={filteredRollingStockList}
+          setIsLoading={setIsLoading}
+        />
+        {displayList()}
         {isAdding && (
           <div className="d-flex flex-column pl-0 rollingstock-editor-form-container mb-3">
             <RollingStockEditorForm
@@ -49,63 +133,6 @@ export default function RollingStockEditor({ rollingStocks }: RollingStockEditor
               setIsAdding={setIsAdding}
             />
           </div>
-        )}
-        {rollingStocks.length > 0 ? (
-          rollingStocks.map((data) => (
-            <div className="rollingstock-editor-list" key={data.id}>
-              <div className="d-flex">
-                <div
-                  role="button"
-                  tabIndex={-1}
-                  className="d-flex align-self-start rollingstock-elements w-100"
-                  onClick={() => {
-                    setIsEditing(false);
-                    setIsAdding(false);
-                  }}
-                >
-                  <RollingStockCard
-                    isOnEditMode
-                    rollingStock={data}
-                    noCardSelected={openedRollingStockCardId === undefined}
-                    isOpen={data.id === openedRollingStockCardId}
-                    setOpenedRollingStockCardId={setOpenedRollingStockCardId}
-                    ref2scroll={rollingStockID === data.id ? ref2scroll : undefined}
-                  />
-                </div>
-                {data.id === openedRollingStockCardId && (
-                  <div className="align-self-start">
-                    <RollingStockEditorButtons
-                      isCondensed
-                      rollingStock={data}
-                      setIsEditing={setIsEditing}
-                      isRollingStockLocked={data.locked as boolean}
-                    />
-                  </div>
-                )}
-              </div>
-              {openedRollingStockCardId === data.id && (
-                <div className="d-flex flex-column pl-0 rollingstock-editor-form-container mb-3">
-                  {(selectedRollingStock || isEditing) &&
-                    ((selectedRollingStock && !isEditing && (
-                      <RollingStockEditorCard
-                        isEditing={isEditing}
-                        data={selectedRollingStock}
-                        setIsEditing={setIsEditing}
-                      />
-                    )) ||
-                      (isEditing && (
-                        <RollingStockEditorForm
-                          rollingStockData={selectedRollingStock}
-                          setIsEditing={setIsEditing}
-                          setIsAdding={setIsAdding}
-                        />
-                      )))}
-                </div>
-              )}
-            </div>
-          ))
-        ) : (
-          <Loader msg={t('rollingstock:waitingLoader')} />
         )}
       </div>
       {!selectedRollingStock && (
