+++ front/src/common/RollingStockSelector/RollingStockModal.tsx
@@ -2,8 +2,6 @@ import React, { useState, useEffect, useContext, useMemo, MutableRefObject } fro
 import { useSelector, useDispatch } from 'react-redux';
 import { setFailure } from 'reducers/main';
 import { useTranslation } from 'react-i18next';
-import { BsLightningFill } from 'react-icons/bs';
-import { MdLocalGasStation } from 'react-icons/md';
 import { isEmpty } from 'lodash';
 
 import { LightRollingStock } from 'common/api/osrdEditoastApi';
@@ -11,68 +9,11 @@ import { enhancedEditoastApi } from 'common/api/enhancedEditoastApi';
 import { RootState } from 'reducers';
 import { getRollingStockID } from 'reducers/osrdconf/selectors';
 import Loader from 'common/Loader';
-import CheckboxRadioSNCF from 'common/BootstrapSNCF/CheckboxRadioSNCF';
-import InputSNCF from 'common/BootstrapSNCF/InputSNCF';
 import ModalBodySNCF from 'common/BootstrapSNCF/ModalSNCF/ModalBodySNCF';
 import { ModalContext } from 'common/BootstrapSNCF/ModalSNCF/ModalProvider';
 import RollingStockEmpty from './RollingStockEmpty';
 import RollingStockCard from './RollingStockCard';
-
-interface Filters {
-  text: string;
-  elec: boolean;
-  thermal: boolean;
-}
-function rollingStockPassesSearchedStringFilter(
-  name: string,
-  metadata: LightRollingStock['metadata'],
-  filters: Filters
-) {
-  if (!filters.text) {
-    return true;
-  }
-  function includesSearchedString(str: string) {
-    return str && str.toLowerCase().includes(filters.text);
-  }
-  return [
-    name,
-    metadata.detail,
-    metadata.reference,
-    metadata.series,
-    metadata.type,
-    metadata.grouping,
-  ].some(includesSearchedString);
-}
-
-export function rollingStockPassesEnergeticModeFilters(
-  modes: LightRollingStock['effort_curves']['modes'],
-  { elec, thermal }: Filters
-) {
-  if (elec || thermal) {
-    const effortCurveModes = Object.values(modes).map(({ is_electric: isElec }) => isElec);
-    const hasAnElectricMode = effortCurveModes.includes(true);
-    const hasAThermalMode = effortCurveModes.includes(false);
-    if ((elec && !hasAnElectricMode) || (thermal && !hasAThermalMode)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-function filterRollingStocks(rollingStocks: LightRollingStock[], filters: Filters) {
-  return rollingStocks?.filter(({ name, metadata, effort_curves: effortCurves }) => {
-    const passSearchedStringFilter = rollingStockPassesSearchedStringFilter(
-      name,
-      metadata,
-      filters
-    );
-    const passEnergeticModesFilter = rollingStockPassesEnergeticModeFilters(
-      effortCurves.modes,
-      filters
-    );
-    return passSearchedStringFilter && passEnergeticModesFilter;
-  });
-}
+import SearchRollingStock from './SearchRollingStock';
 
 interface RollingStockModal {
   ref2scroll: MutableRefObject<HTMLDivElement | null>;
@@ -83,11 +24,6 @@ function RollingStockModal({ ref2scroll }: RollingStockModal) {
   const darkmode = useSelector((state: RootState) => state.main.darkmode);
   const rollingStockID = useSelector(getRollingStockID);
   const { t } = useTranslation(['translation', 'rollingstock']);
-  const [filters, setFilters] = useState({
-    text: '',
-    elec: false,
-    thermal: false,
-  });
   const [isLoading, setIsLoading] = useState(false);
   const [openRollingStockCardId, setOpenRollingStockCardId] = useState(rollingStockID);
   const { closeModal } = useContext(ModalContext);
@@ -98,43 +34,13 @@ function RollingStockModal({ ref2scroll }: RollingStockModal) {
 
   const {
     data: { results: rollingStocks } = { results: [] },
-    isSuccess,
     isError,
     error,
   } = enhancedEditoastApi.useGetLightRollingStockQuery({
     pageSize: 1000,
   });
-  const [filteredRollingStockList, setFilteredRollingStockList] = useState<LightRollingStock[]>(
-    () => filterRollingStocks(rollingStocks, filters)
-  );
-
-  function handleRollingStockLoaded() {
-    const newFilteredRollingStock = filterRollingStocks(rollingStocks, filters);
-    setFilteredRollingStockList(newFilteredRollingStock);
-  }
-
-  const updateSearch = () => {
-    if (filters.text !== '' || filters.elec !== false || filters.thermal !== false) {
-      setOpenRollingStockCardId(undefined);
-    } else {
-      setOpenRollingStockCardId(rollingStockID);
-    }
-    const newFilteredRollingStock = filterRollingStocks(rollingStocks, filters);
-    setTimeout(() => {
-      setFilteredRollingStockList(newFilteredRollingStock);
-      setIsLoading(false);
-    }, 0);
-  };
-
-  const searchMateriel = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFilters({ ...filters, text: e.target.value.toLowerCase() });
-    setIsLoading(true);
-  };
-
-  const toggleFilter = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setFilters({ ...filters, [e.target.name]: !filters[e.target.name as 'elec' | 'thermal'] });
-    setIsLoading(true);
-  };
+  const [filteredRollingStockList, setFilteredRollingStockList] =
+    useState<LightRollingStock[]>(rollingStocks);
 
   useEffect(() => {
     if (openRollingStockCardId !== undefined) {
@@ -145,10 +51,6 @@ function RollingStockModal({ ref2scroll }: RollingStockModal) {
       }, 1000);
     }
   }, [ref2scroll.current]);
-
-  useEffect(() => {
-    handleRollingStockLoaded();
-  }, [isSuccess, rollingStocks]);
 
   useEffect(() => {
     if (isError && error && 'status' in error) {
@@ -164,17 +66,10 @@ function RollingStockModal({ ref2scroll }: RollingStockModal) {
     }
   }, [isError]);
 
-  useEffect(() => {
-    if (rollingStocks && rollingStocks.length !== 0) {
-      updateSearch();
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [filters]);
-
   const listOfRollingStocks = useMemo(
     () =>
       filteredRollingStockList.length > 0 ? (
-        filteredRollingStockList.map((item) => (
+        filteredRollingStockList.map((item: LightRollingStock) => (
           <RollingStockCard
             rollingStock={item}
             key={item.id}
@@ -207,62 +102,14 @@ function RollingStockModal({ ref2scroll }: RollingStockModal) {
           <button type="button" className="close" aria-label="Close" onClick={closeModal}>
             <span aria-hidden="true">&times;</span>
           </button>
-          <div className="row no-gutters">
-            <div className="col-md-4 mb-3">
-              <InputSNCF
-                id="searchfilter"
-                type="text"
-                onChange={searchMateriel}
-                placeholder={t('translation:common.search')}
-                noMargin
-                unit={<i className="icons-search" />}
-                sm
-              />
-            </div>
-            <div className="col-md-5 ml-2 mb-3 d-flex align-items-center">
-              <div className="mr-4">
-                <CheckboxRadioSNCF
-                  onChange={toggleFilter}
-                  name="elec"
-                  id="elec"
-                  label={
-                    <span className="text-nowrap">
-                      <span className="text-primary mr-1">
-                        <BsLightningFill />
-                      </span>
-                      {t('rollingstock:electric')}
-                    </span>
-                  }
-                  type="checkbox"
-                  checked={filters.elec}
-                />
-              </div>
-              <div>
-                <CheckboxRadioSNCF
-                  onChange={toggleFilter}
-                  name="thermal"
-                  id="thermal"
-                  label={
-                    <span className="text-nowrap">
-                      <span className="text-pink mr-1">
-                        <MdLocalGasStation />
-                      </span>
-                      {t('rollingstock:thermal')}
-                    </span>
-                  }
-                  type="checkbox"
-                  checked={filters.thermal}
-                />
-              </div>
-            </div>
-            <div className="col-md-2 mt-1 ml-auto">
-              <small className="">
-                {filteredRollingStockList.length > 0
-                  ? `${filteredRollingStockList.length} ${t('rollingstock:resultsFound')}`
-                  : t('rollingstock:noResultFound')}
-              </small>
-            </div>
-          </div>
+          <SearchRollingStock
+            rollingStocks={rollingStocks}
+            rollingStockID={rollingStockID}
+            setOpenedRollingStockCardId={setOpenRollingStockCardId}
+            setFilteredRollingStockList={setFilteredRollingStockList}
+            filteredRollingStockList={filteredRollingStockList}
+            setIsLoading={setIsLoading}
+          />
         </div>
         <div className="rollingstock-search-list">{displayList()}</div>
       </div>
